+++ buildings/gui/new_capture_source_area.py
@@ -6,7 +6,7 @@ from functools import partial
 from PyQt4 import uic
 from PyQt4.QtCore import pyqtSlot
 from PyQt4.QtGui import QFrame, QToolButton, QTableWidgetItem, QHeaderView, QAbstractItemView
-from qgis.core import QgsFeatureRequest, QgsGeometry, QgsMapLayer, QgsMapLayerRegistry
+from qgis.core import QGis, QgsFeatureRequest, QgsGeometry, QgsMapLayer, QgsMapLayerRegistry
 from qgis.gui import QgsMessageBar
 from qgis.utils import iface
 
@@ -122,7 +122,7 @@ class NewCaptureSourceArea(QFrame, FORM_CLASS):
         elif len(selection) == 1:
             new_geometry = selection[0].geometry()
             # error pops up if geometry type is not polygon or multipolygon
-            if new_geometry.type() not in [2, 4]:
+            if new_geometry.wkbType() not in [QGis.WKBPolygon, QGis.WKBMultiPolygon]:
                 self.error_dialog = ErrorDialog()
                 self.error_dialog.fill_report(
                     '\n -------------------- WRONG GEOMETRY TYPE ------'
@@ -134,7 +134,7 @@ class NewCaptureSourceArea(QFrame, FORM_CLASS):
             self.le_external_id.setEnabled(True)
             self.le_area_title.setEnabled(True)
             # convert to correct format
-            if new_geometry.type() == 2:
+            if new_geometry.wkbType() == QGis.WKBPolygon:
                 new_geometry = QgsGeometry.fromMultiPolygon([new_geometry.asPolygon()])
             wkt = new_geometry.exportToWkt()
             sql = general_select.convert_geometry
@@ -237,7 +237,7 @@ class NewCaptureSourceArea(QFrame, FORM_CLASS):
         new_feature = next(self.capture_source_area.getFeatures(request))
         new_geometry = new_feature.geometry()
         # convert to correct format
-        if new_geometry.type() == 2:
+        if new_geometry.wkbType() == QGis.WKBPolygon:
             new_geometry = QgsGeometry.fromMultiPolygon([new_geometry.asPolygon()])
         wkt = new_geometry.exportToWkt()
         sql = general_select.convert_geometry
@@ -270,7 +270,7 @@ class NewCaptureSourceArea(QFrame, FORM_CLASS):
            @type  geom:        qgis.core.QgsGeometry
         """
         if qgsfId in self.added_building_ids:
-            if geom.type() == 2:
+            if geom.wkbType() == QGis.WKBPolygon:
                 geom = QgsGeometry.fromMultiPolygon([geom.asPolygon()])
             wkt = geom.exportToWkt()
             if not wkt:
@@ -334,6 +334,7 @@ class NewCaptureSourceArea(QFrame, FORM_CLASS):
         self.mcb_selection_layer.setDisabled(True)
         self.l_wrong_projection.setText('')
 
+        iface.setActiveLayer(self.capture_source_area)
         self.capture_source_area.geometryChanged.disconnect(self.creator_geometry_changed)
         iface.actionCancelEdits().trigger()
         self.capture_source_area.geometryChanged.connect(self.creator_geometry_changed)